Update CalculateConversion
Se deja en double

+++ src/com/currencyconverter/calculos/CalculateConversion.java
@@ -3,10 +3,10 @@ package com.currencyconverter.calculos;
 import com.currencyconverter.models.Currency;
 
 public class CalculateConversion {
-    private long conversionResult;
+    private double conversionResult;
 
     public void setConversionResult(Currency currency) {
-        this.conversionResult = (long) (currency.getValueToConvert() * currency.getConversionRate());
+        this.conversionResult = currency.getValueToConvert() * currency.getConversionRate();
         currency.setConversionResult(this.conversionResult);
     }
 }
